Remove comments
Remove a bunch of unnecessary comments, the code should look clear and
not filled with comments. If something is needed we can check the
previous lines using git history.

Also there should not be more than one empty line between sentences, it
adds unnecessary empty lines and make the code hard to read.