Add helpers enums
These enums will help to know if the current actor is either part of
the map, the player of any enemies, also will help to know the enemy
type and will help to handle the difficulty for when we add new levels
(i.e. showing more enemies with higher HP in the level).